+++ test/Rhetos.Core.Test/DatabaseGenerator/ConceptApplicationRepositoryTest.cs
@@ -195,7 +195,9 @@ namespace Rhetos.DatabaseGenerator.Test
                 throw new NotImplementedException();
             }
 
-            public int GetTransactionCount() => 1;
+            public int GetTransactionInitialState() => 1;
+
+            public string CheckTransactionState(int initialState) => null;
 
             public void GetDbLock(IEnumerable<string> resources, bool wait = true)
             {